% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{aabc.cv.func}
\alias{aabc.cv.func}
\title{Internal function used in xselect.glmnet.}
\usage{
aabc.cv.func(data, yvar, xvars, trtvar, censorvar, type)
}
\arguments{
\item{data}{the original dataset.}

\item{yvar}{the original outcome variable name.}

\item{xvars}{the name of predictors.}

\item{trtvar}{treatment variable name.}

\item{censorvar}{the name of censor variable.}

\item{type}{"b" for binary outcome or "s" for time to event outcome.}
}
\value{
cross validated aABC for selected predictors.
}
\description{
Internal function used in xselect.glmnet.
}
