% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{AUC.emp}
\alias{AUC.emp}
\title{Function for AUC when input is X and Y.}
\usage{
AUC.emp(X, Y)
}
\arguments{
\item{X}{nondiseased sample}

\item{Y}{diseased sample}
}
\value{
AUC for given dieased and non-diseased sample
}
\description{
Function for AUC when input is X and Y.
}
