% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{aucabc.func}
\alias{aucabc.func}
\title{Function for AUC/ABC used in cv.xselect.}
\usage{
aucabc.func(score, data, type, trtvar, censorvar, yvar, xvars)
}
\arguments{
\item{score}{composite score from cv.xselect.}

\item{data}{the training dataset.}

\item{type}{type of response.}

\item{trtvar}{treatment variable name.}

\item{censorvar}{censoring variable name.}

\item{yvar}{response variable name.}

\item{xvars}{a vector of predictor names.}
}
\value{
AUC/ABC for each score from cv.xselect.
}
\description{
Function for AUC/ABC used in cv.xselect.
}
