% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{betahat.fold.func}
\alias{betahat.fold.func}
\title{beta estimates for each candidate predictor set.}
\usage{
betahat.fold.func(
  yvar,
  xvars,
  censorvar,
  trtvar,
  trtref,
  data,
  type,
  pmax,
  family,
  nfolds = 5
)
}
\arguments{
\item{yvar}{response variable name.}

\item{xvars}{a vector of predictor variable names.}

\item{censorvar}{censoring variable name.}

\item{trtvar}{treatment variable name.}

\item{trtref}{treatment reference code.}

\item{data}{data frame for train data.}

\item{type}{"b" binary; "s" survival.}

\item{pmax}{maximum number of variables ever to be nonzero as required in glmnet function.}

\item{family}{response type.}

\item{nfolds}{number of CV folder used in cv.glmnet.}
}
\value{
the coefficient estimates for each fold in internal CV.
}
\description{
beta estimates for each candidate predictor set.
}
