% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{betahatX}
\alias{betahatX}
\title{optAUC estimates for predictive case.}
\usage{
betahatX(
  yvar,
  xvars,
  censorvar = NULL,
  trtvar,
  trtref,
  data,
  type = "b",
  xselect = FALSE,
  cv.iter.xselect = 20,
  k.fold.xselect = 5,
  method.xselect = "CV"
)
}
\arguments{
\item{yvar}{outcome variable (column) name for response variable assuming data is provided.}

\item{xvars}{vector of variable names for predictors (covariates) assuming data is provided.}

\item{censorvar}{name of censoring variable (1: event; 0: censor), default = NULL.}

\item{trtvar}{traetment variable name.}

\item{trtref}{coding (in the column of treatment) for treatment arm.}

\item{data}{data frame for training dataset.}

\item{type}{type of response variable: "b" binary; "s" survival.}

\item{xselect}{a logical flag. TRUE: use Cross Validated ABC for variable selection; FALSE: use all predictors to calculate the composite score. Default = FALSE.}

\item{cv.iter.xselect}{the number of iterations of cross-validation used for MCglmnet.}

\item{k.fold.xselect}{the number of folds required for cross-validated ABC.}

\item{method.xselect}{the method used for variable selection: CV (default) cross validated ABC.}
}
\value{
This function returns a list of the following:
\item{beta}{the estimated linear coefficients.}
\item{bias}{the HIC for the selected model.}
}
\description{
Fit optAUC to the data in predictive case. Estimates along with the model bias correction are returned.
}
