% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{C.index}
\alias{C.index}
\title{This function is for calculating C index.}
\usage{
C.index(yvar, score, censorvar, data)
}
\arguments{
\item{yvar}{is the outcome variable name}

\item{score}{is the composite score}

\item{censorvar}{is the censoring variable name}

\item{data}{is the data frame used}
}
\value{
C index for the survival outcome.
}
\description{
This function is for calculating C index.
}
\author{
Xin Huang, Yan Sun, Lu Tian, Saptarshi Chatterjee, Viswanath Devanarayan
}
