% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{cv.optAUCX.fit}
\alias{cv.optAUCX.fit}
\title{CV function for optAUCX.fit}
\usage{
cv.optAUCX.fit(
  data,
  yvar,
  yvar.ref = 1,
  xvars,
  censorvar = NULL,
  trtvar = NULL,
  type = "b",
  trtref = 1,
  des.res = "larger",
  method.subgrp = "BATTing",
  scale = TRUE,
  xselect = TRUE,
  cv.iter.xselect = 20,
  k.fold.xselect = 5,
  method.xselect = "CV",
  pre.filter = "opt",
  filter.method = NULL,
  n.boot = 10,
  cv.iter.eval = 20,
  k.fold.eval = 5,
  max.iter = 500
)
}
\arguments{
\item{data}{the input data frame.}

\item{yvar}{response variable name.}

\item{yvar.ref}{response variable reference.}

\item{xvars}{vector of predictor variable names.}

\item{censorvar}{censoring variable name only for TTE data , censor=0,event=1)  - default = NULL.}

\item{trtvar}{treatment variable name (for predictive signature). Set trt.vec to NULL for prognostic signature.}

\item{type}{type of response variable: "s" survival; "b" binary (default).}

\item{trtref}{coding (in the column of trtvar) for treatment arm.}

\item{des.res}{the desired response. "larger": prefer larger response (default) "smaller": prefer smaller response.}

\item{method.subgrp}{method for subgroup identification: "BATTing" or "Youden".}

\item{scale}{logical variable indicating if scaling is required.}

\item{xselect}{a logical flag. TRUE: use LASSO for variable selection; FALSE: use all predictors to calculate the composite score. Default = TRUE.}

\item{cv.iter.xselect}{the number of iterations of cross-validation used for xselect.}

\item{k.fold.xselect}{the number of folds of cross-validation used for xselect}

\item{method.xselect}{the method used for variable selection: CV (default) cross validated AUC/ABC; "aAUC" approximate AUC (only for prognostic case).}

\item{pre.filter}{it is either "opt" or an integer between 1 and total number of predictors. Default is "opt", no prefiltering conducted, optimized number of predictors selected; An integer: min(opt, integer) of predictors selected}

\item{filter.method}{method for filtering: NULL, no prefiltering, "univariate", univaraite filtering; "glmnet", glmnet filtering, "unicart": univariate rpart filtering for prognostic case.}

\item{n.boot}{number of bootstrap for batting or Youden procedure.}

\item{cv.iter.eval}{the number of iterations for outer cross-validation.}

\item{k.fold.eval}{the number of folds for outer cross-validation.}

\item{max.iter}{Maximum number of successful iterations required for CV.}
}
\value{
a list containing with following entries:
\item{stats.summary}{Summary of performance statistics.}
\item{pred.classes}{Data frame containing the predictive clases (TRUE/FALSE) for each iteration.}
\item{folds}{Data frame containing the fold indices (index of the fold for each row) for each iteration.}
\item{sig.list}{List of length cv.iter * k.fold containing the signature generated at each of the  k folds, for all iterations.}
\item{error.log}{List of any error messages that are returned at an iteration.}
\item{interplot}{Treatment*subgroup interaction plot for predictive case}
}
\description{
CV function for optAUCX.fit
}
