% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{cv.xselect}
\alias{cv.xselect}
\title{Function for internal CV to estimate number of candidate predictors required for composite score.}
\usage{
cv.xselect(
  data,
  yvar,
  xvars,
  censorvar,
  trtvar,
  type,
  trtref,
  cv.iter.xselect,
  k.fold.xselect,
  pmax,
  family
)
}
\arguments{
\item{data}{the dataset used for training.}

\item{yvar}{the original outcome variable name.}

\item{xvars}{the name of predictors.}

\item{censorvar}{the name of censor variable.}

\item{type}{"b" for binary outcome or "s" for time to event outcome.}

\item{trtref}{the coding (in the column of treatment) for treatment arm.}

\item{cv.iter.xselect}{number of iterations for internal cross-validation.}

\item{k.fold.xselect}{number of folds for internal cross-validation.}
}
\value{
returns median and se of cross-validated AUC/ABC and across internal cv iterations (required for 1se rule).
}
\description{
Function for internal CV to estimate number of candidate predictors required for composite score.
}
