% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{grad.sub}
\alias{grad.sub}
\title{function of grad_square in the GCV}
\usage{
grad.sub(z, beta)
}
\arguments{
\item{z}{(m x n) x p data matrix as defined in betahat/betahatX function.}

\item{beta}{coefficient estimates from betahat/betahatX function.}
}
\value{
grad_square in the GCV
}
\description{
function of grad_square in the GCV
}
