% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{gradsqr}
\alias{gradsqr}
\title{Internal function for HIC calculation}
\usage{
gradsqr(beta, X, Y, Z0, w = 1, weights = 1)
}
\arguments{
\item{beta}{coefficient estimates from betahat/betahatX function.}

\item{X}{m x p data matrix for m non-diseased subjects with p markers.}

\item{Y}{n x p data matrix for n diseased subjects with p markers.}

\item{Z0}{(m x n) x p data matrix as defined in betahat/betahatX function.}

\item{w}{inverse probability weighting for missing data (default is 1).}

\item{weights}{how much weight each predictor should get.}
}
\value{
gradient square for the GCV.
}
\description{
Internal function for HIC calculation
}
