% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{hessAUC}
\alias{hessAUC}
\title{function for Hessian matrix of AUC.}
\usage{
hessAUC(beta, Z, w = 1)
}
\arguments{
\item{beta}{coefficient estimates from betahat/betahatX function.}

\item{Z}{(m x n) x p data matrix as defined in betahat/betahatX function.}

\item{w}{inverse probability weighting for missing data (default is 1).}
}
\value{
Hessian matrix of AUC.
}
\description{
function for Hessian matrix of AUC.
}
