% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{hessAUC.sub}
\alias{hessAUC.sub}
\title{Internal function for hessAUC.}
\usage{
hessAUC.sub(z, beta)
}
\arguments{
\item{z}{(m x n) x p data matrix as defined in betahat/betahatX function.}

\item{beta}{coefficient estimates from betahat/betahatX function.}
}
\value{
Hessian matrix components.
}
\description{
Internal function for hessAUC.
}
