% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{HIC}
\alias{HIC}
\title{Function for HIC calculation.}
\usage{
HIC(beta, X, Y, Z, weights)
}
\arguments{
\item{beta}{coefficient estimates from betahat/betahatX function.}

\item{X}{m x p data matrix for m non-diseased subjects with p markers.}

\item{Y}{m x p data matrix for m diseased subjects with p markers.}

\item{Z}{(m x n) x p data matrix as defined in betahat/betahatX function.}

\item{weights}{weight of the variables}
}
\value{
A numeric value with corresponding HIC.
}
\description{
Function for HIC calculation.
}
