% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{optAUCX.fit}
\alias{optAUCX.fit}
\title{Fit function for optAUCX, fit linear composite and find threshold for subgroup}
\usage{
optAUCX.fit(
  data,
  yvar,
  yvar.ref = 1,
  xvars,
  censorvar = NULL,
  trtvar = NULL,
  type = "b",
  trtref = 1,
  des.res = "larger",
  method.subgrp = "Youden",
  scale = TRUE,
  xselect = TRUE,
  cv.iter.xselect = 20,
  k.fold.xselect = 5,
  method.xselect = "CV",
  pre.filter = "opt",
  filter.method = NULL,
  n.boot = 10,
  min.sigp.prcnt = 0.2
)
}
\arguments{
\item{data}{the input data frame}

\item{yvar}{response variable name}

\item{yvar.ref}{response variable reference}

\item{xvars}{vector of predictor variable names}

\item{censorvar}{censoring variable name only for TTE data , censor=0,event=1)  - default = NULL.}

\item{trtvar}{treatment variable name (for predictive signature). Set trt.vec to NULL for prognostic signature.}

\item{type}{type of response variable: "s" survival; "b" binary (default).}

\item{trtref}{coding (in the column of trtvar) for treatment arm.}

\item{des.res}{the desired response. "larger": prefer larger response (default) "smaller": prefer smaller response}

\item{method.subgrp}{method for subgroup identification: "BATTing" or "Youden"}

\item{scale}{logical variable indicating if scaling is required}

\item{xselect}{it is a logical flag. TRUE: use LASSO for variable selection; FALSE: use all predictors to calculate the composite score. Default = TRUE}

\item{cv.iter.xselect}{it is the number of iterations of cross-validation used for xselect}

\item{k.fold.xselect}{it is the number of folds of cross-validation used for xselect}

\item{method.xselect}{it is the method used for variable selection: CV (default) cross validated AUC/ABC; "aAUC" approximate AUC (only for prognostic case)}

\item{pre.filter}{it is either "opt" or an integer between 1 and total number of predictors. Default is "opt", no prefiltering conducted, optimized number of predictors selected; An integer: min(opt, integer) of predictors selected}

\item{filter.method}{method for filtering: NULL, no prefiltering, "univariate", univaraite filtering; "glmnet", glmnet filtering, "unicart": univariate rpart filtering for prognostic case}

\item{n.boot}{number of bootstrap for batting or Youden procedure}

\item{min.sigp.prcnt}{desired proportion of signature positive group size for a given cutoff.}
}
\value{
This function returns a list containing following entries
\item{optAUC}{output entries as returned by optAUCX function.}
\item{subgrp}{a list containing prediction rule based on optAUC score, pvalue and group summary statistics of each subgroup and interaction plot (only for predictive case)}
}
\description{
Fit function for optAUCX, fit linear composite and find threshold for subgroup
}
\examples{
library(optAUCX)
library(MASS)
library(stringr)

}
