% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{optAUCX}
\alias{optAUCX}
\title{Optimal Combination of biomarkers for prognostic and predictive signature development Based on total AUC.}
\usage{
optAUCX(
  outcome,
  predictor,
  censorvar = NULL,
  treatment = NULL,
  type = "b",
  data = NULL,
  resp.ref = 1,
  trt.ref = 1,
  xselect = FALSE,
  scale = TRUE,
  n.boot = 10,
  cv.iter.xselect = 20,
  k.fold.xselect = 5,
  method.xselect = "CV",
  pre.filter = NULL,
  filter.method = NULL
)
}
\arguments{
\item{outcome}{variable (column) name for response variable if data is provided; otherwise, a vector of responses.}

\item{predictor}{vector of variable names for predictors (covariates) if data is provided; otherwise, a data matrix of predictors.}

\item{censorvar}{name of censoring variable (1: event; 0: censor), default = NULL.}

\item{treatment}{variable name for treatment variable if data is provided; otherwise, a vector of treatment variables. default = NULL (prognostic signature).}

\item{type}{type of response variable: "b" binary; "s" survival.}

\item{data}{data frame for training dataset.}

\item{resp.ref}{coding (in the column of outcome) for responder, used for binary outcome.}

\item{trt.ref}{coding (in the column of treatment) for treatment arm.}

\item{xselect}{a logical flag. TRUE: use LASSO for variable selection; FALSE: use all predictors to calculate the composite score. Default = FALSE.}

\item{scale}{a logical flag indicating whether standardization is to be performed to the dataset before the combination, default is TRUE.}

\item{n.boot}{the number of bootstrap for the MCglmnet procedure.}

\item{cv.iter.xselect}{the number of iterations for internal cross-validation used for xselect.}

\item{k.fold.xselect}{the number of folds for internal cross-validation used for xselect.}

\item{method.xselect}{the method used for variable selection: "CV" (default) cross validated AUC/ABC; "aAUC" approximate AUC (only for prognostic case).}

\item{pre.filter}{NULL, no prefiltering conducted;"opt", optimized number of predictors selected; An integer: min(opt, integer) of predictors selected.}

\item{filter.method}{NULL, no prefiltering, "univariate", univaraite filtering; "glmnet", glmnet filtering, "unicart": univariate rpart filtering for prognostic case.}
}
\value{
a list containing following entries:
\item{beta}{linear coefficient estimates.}
\item{data}{the dataset with composite score.}
\item{AUC}{the apparent AUC (C-index if survival outcome) calculated from the estimated beta (prognostic case only).}
\item{trt.AUC}{the apparent AUC (C-index if survival outcome) calculated from the estimated beta for the treatment arm (predictive case only).}
\item{ctl.AUC}{the apparent AUC (C-index if survival outcome) calculated from the estimated beta for the control arm (predictive case only).}
\item{ABC}{the area between the apparent ROCs (C-index if survival outcome) calculated from the estimated beta for the control arm (predictive case only).}

\item{sd}{a numeric vector of standard deviations of the columns of the training predictors.}
\item{mean}{a numeric vector of mean of the columns of the training predictors.}
\item{aAUC}{the approximate AUC correcting the bias of the model (prognostic case only).}
\item{aABC}{the approximate ABC correcting the bias of the model (predictive case only).}
}
\description{
Searching for optimal linear combination of multiple biomarkers that (1) for prognostic signature development, maximizes the total area under the receiver operating characteristic (ROC) curves (for binary outcome)/C-index (for survival outcome); (2) for predictive signature development, maximizes the area between ROC curves (for binary outcome)/C-index (for survival outcome) between treatment and control arm.
}
