% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{optAUCX.wrapper}
\alias{optAUCX.wrapper}
\title{The wrapper function for optAUCX, to be passed to kfold.cv.}
\usage{
optAUCX.wrapper(data, args)
}
\arguments{
\item{data}{data frame equal to cbind(y, x), where y and x are inputs to optAUCX.}

\item{args}{list containing all other input arguments to optAUCX. except for x and y. Also contains xvars=names(x) and yvar=names(y).}
}
\value{
prediction rule returned by optAUCX.
}
\description{
The wrapper function for optAUCX, to be passed to kfold.cv.
}
