% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{pred.optAUCX.CV}
\alias{pred.optAUCX.CV}
\title{Prediction function for optAUC CV.}
\usage{
pred.optAUCX.CV(data, predict.rule, args)
}
\arguments{
\item{data}{data frame equal to cbind(y, x), where y and x are inputs to optAUCX.}

\item{predict.rule}{prediction rule returned by optAUCX.fit.}

\item{args}{prediction rule arguments.}
}
\value{
Data frame containing two columns: y and pred.class, a logical vector indicating the prediction for each row of data.
}
\description{
Predict outcome from optAUCX fit for cross-validation data given trained model in optAUCX.fit. This is intended to be used with kfold.cv.
}
