% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{pred.optAUCX}
\alias{pred.optAUCX}
\title{Prediction function for optAUCX.fit}
\usage{
pred.optAUCX(x, res.optAUC.fit)
}
\arguments{
\item{x}{input predictors matrix.}

\item{res.optAUC.fit}{Prediction rule returned by optAUCX.fit.}
}
\value{
a logical vector indicating the prediction for each row of data.
}
\description{
Prediction function for optAUCX.fit
}
