% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{rank.list.func}
\alias{rank.list.func}
\title{Internal function used in betahat.fold.func}
\usage{
rank.list.func(betamat, xvars)
}
\arguments{
\item{betamat}{betahat matrix from cv.glmnet.}

\item{xvars}{a vector of predictor variables.}
}
\value{
order of each set of predictors as returned by cv.glmnet.
}
\description{
Internal function used in betahat.fold.func
}
