% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{seq.find.cutoff.Youden}
\alias{seq.find.cutoff.Youden}
\title{Find cutoff for predictive case using Youden score.}
\usage{
\method{seq}{find.cutoff.Youden}(
  data,
  yvar,
  censorvar,
  xvar,
  trtvar,
  trtref,
  type,
  dir,
  nsubj,
  min.sigp.prcnt
)
}
\arguments{
\item{data}{input data frame.}

\item{yvar}{response variable name.}

\item{censorvar}{censoring variable name only for TTE data , censor=0,event=1)  - default = NULL.}

\item{xvar}{composite score from optAUC.}

\item{trtvar}{treatment variable name (for predictive signature). Set trt.vec to NULL for prognostic signature.}

\item{trtref}{coding (in the column of trtvar) for treatment arm.}

\item{type}{type of response variable: "s" survival; "b" binary (default).}

\item{dir}{direction of cut.}

\item{nsubj}{number of subjects.}

\item{min.sigp.prcnt}{desired proportion of signature positive group size for a given cutoff.}
}
\value{
The cutpoint corresponding to maximum Youden score.
}
\description{
Find cutoff for predictive case using Youden score.
}
