% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{seq.find.score.Youden}
\alias{seq.find.score.Youden}
\title{Youden score for a specific cutpoint}
\usage{
\method{seq}{find.score.Youden}(data, yvar, xvar, dir, cutoff, min.sigp.prcnt)
}
\arguments{
\item{data}{input data frame.}

\item{yvar}{response variable name.}

\item{xvar}{composite score from optAUC.}

\item{dir}{direction of cut.}

\item{cutoff}{a specific cutpoint for which the score needs to be computed.}

\item{min.sigp.prcnt}{desired proportion of signature positive group size for a given cutoff.}
}
\value{
Youden score for the cutpoint specified.
}
\description{
Youden score for a specific cutpoint
}
