% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalAUCX_functions_surv.R
\name{xselect.glmnet}
\alias{xselect.glmnet}
\title{Internal function for selecting candidate predictors for composite score.}
\usage{
xselect.glmnet(
  data,
  type,
  yvar,
  xvars,
  data.orig,
  yvar.orig,
  censorvar = NULL,
  trtvar = NULL,
  trtref = 1,
  n.boot = 50,
  cv.iter.xselect = 20,
  k.fold.xselect = 5,
  method.xselect = "CV"
)
}
\arguments{
\item{data}{the pseudo dataset as used in betahat or betahatX function.}

\item{type}{type of response variable: "b" binary; "s" survival.}

\item{yvar}{simulated binary variable name used in betahat or betahatX function.}

\item{xvars}{vector of variable names for predictors (covariates) assuming data is provided.}

\item{data.orig}{the original dataset used for training.}

\item{yvar.orig}{the original outcome variable (column) name.}

\item{censorvar}{name of censoring variable (1: event; 0: censor), default = NULL.}

\item{trtvar}{treatment variable name.}

\item{trtref}{the coding (in the column of treatment) for treatment arm.}

\item{n.boot}{the number of bootstrap for the MCglmnet procedure.}

\item{cv.iter.xselect}{the number of iterations of cross-validation used for MCglmnet.}

\item{k.fold.xselect}{the number of folds required for cross-validated AUC or ABC.}

\item{method.xselect}{the method used for variable selection: CV (default) cross validated AUC or ABC; "aAUC" approximate AUC (Only for prognostic case).}
}
\value{
the variables selected for composite score.
}
\description{
Internal function for selecting candidate predictors for composite score.
}
