% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optAUCx.lib.updated.R
\name{Youden.func}
\alias{Youden.func}
\title{Function for obtaining cutpoints based on Youden score.}
\usage{
Youden.func(
  data,
  yvar,
  xvar,
  trtvar,
  trtref,
  censorvar,
  type,
  n.boot,
  des.res,
  min.sigp.prcnt
)
}
\arguments{
\item{data}{the input data frame.}

\item{yvar}{response variable name.}

\item{xvar}{composite score from optAUC.}

\item{trtvar}{treatment variable name (for predictive signature). Set trt.vec to NULL for prognostic signature.}

\item{trtref}{coding (in the column of trtvar) for treatment arm.}

\item{censorvar}{censoring variable name only for TTE data , censor=0,event=1)  - default = NULL.}

\item{type}{type of response variable: "s" survival; "b" binary (default).}

\item{n.boot}{number of bootstrap for batting or Youden procedure.}

\item{des.res}{the desired response. "larger": prefer larger response (default) "smaller": prefer smaller response.}

\item{min.sigp.prcnt}{desired proportion of signature positive group size for a given cutoff.}
}
\value{
this function returns a list containing
\item{res}{prediction rule based on optAUC score.}
\item{train.stat}{subgroup p-value and group statistics for training data.}
\item{train.plot}{Treatment-subgroup interaction plot for training data (only for predictive signature).}
}
\description{
Function for obtaining cutpoints based on Youden score.
}
